+++ example/methods_example/track_methods_example.dart
@@ -7,7 +7,7 @@ import 'package:scrobblenaut/lastfm.dart';
 import 'package:scrobblenaut/scrobblenaut.dart';
 import 'package:scrobblenaut/scrobblenaut_helpers.dart';
 
-import 'api_values.dart';
+import '../api_values.dart';
 
 // Just an example of use.
 void main() async {